дизайн и анимации
создал дизайн и анимацию тапа по всем кнопкам. Осталось сделать отображение присутствия на странице. Анимацию отлетающих циферок при тапе

+++ Assets/_Project/Core/Tapping/SoftCurrency.cs
@@ -8,8 +8,8 @@ namespace Core.Tappings
     public class SoftCurrency : MonoBehaviour
     {
         [SerializeField] private GameInitializer _gameInitializer;
-        [SerializeField] private ReactiveField<int> _currentCurrency = new ReactiveField<int>(0);
-        public ReactiveField<int> CurrencyField => _currentCurrency;
+        [SerializeField] private ReactiveField<long> _currentCurrency = new ReactiveField<long>(0);
+        public ReactiveField<long> CurrencyField => _currentCurrency;
         private void Start()
         {
             _currentCurrency.Value = SaveManager.Load("_currentCurrency", 0);
